requirements.txt
mysql-connector-python
flask
base58